Refactor: Update Accept header in NetworkModule
The `Accept` header in `NetworkModule.kt` is updated to include `text/javascript` and `text/plain` in addition to `application/json`.
The `Content-Type` header is removed as it's generally used in requests that include a message body (e.g., POST, PUT) to indicate the media type of the body being sent, which is not applicable here for default request configuration.